requirements.txt
selenium>=4.0.0
requests>=2.0.0
pillow>=8.0.0
